+++ packages/ws-worker/CHANGELOG.md
@@ -1,5 +1,13 @@
 # ws-worker
 
+## 0.1.1
+
+### Patch Changes
+
+- Destroy workers on close
+- Updated dependencies
+  - @openfn/engine-multi@0.1.1
+
 ## 0.1.0
 
 First release of the websocket worker, which handles comm between Lightning and the multi-threaded engine.